Add function to save training data and update environment versions
This commit includes several improvements to the training and evaluation code. We added a function save_pendulum_dataset() to save the training data into an HDF5 file, which could be beneficial for debugging, future analysis, or training iterations. We also updated the Pendulum environment from v0 to v1 and changed the device used for training from GPU to CPU. Furthermore, the code for logging images was modified to use PIL instead of scipy.misc. The progress_bar and no_log options were micro-adjusted to better fit usual use cases. Lastly, the training data is now saved at the end of the training process.